feat(Workspace) : Rewrite workspace in functional component (#67)
