+++ .gitignore
@@ -84,3 +84,5 @@ typings/
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+.yarnrc
