+++ src/components/snackbar/index.js
@@ -1,0 +1,3 @@
+import MuiSnackbar from "./MuiSnackbar";
+
+export default MuiSnackbar;
