+++ src/components/workspace/index.js
@@ -1,0 +1,3 @@
+import Workspace from "./Workspace";
+
+export default Workspace;
